finalize simple generators
