+++ src/main/resources/assets/gregtech/textures/gui/overlay/furnace_overlay_1.png
